requirements.txt
flask>=2.0.0
pytest>=7.0.0
requests>=2.28.0
python-dotenv>=1.0.0
rich>=13.0.0
llama-cpp-python>=0.2.0
